+++ types/node.d.ts
@@ -1,0 +1,5 @@
+declare namespace NodeJS {
+    interface Process {
+        sentry: typeof import('@sentry/node');
+    }
+}
